rename some things for clarification; ultimate goal is to get core.ts and app.ts in the same file
